requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pystray==0.19.5
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
redis==7.3.0
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
six==1.17.0
spotipy==2.26.0
typer==0.24.1
urllib3==2.6.3
